Isolated OWM provider info to fix #22
1) Moved OWMToWeatherConditionMapper into provider/owm_provider.py (from
weather_condition.py) as WeatherConditionMapper.  This allows us to use
a different mapper for each provider.

2) Moved OWMWeatherConditions enum into provider/owm_provider.py, since
it should only be needed there now.  Removed owm.py since that class is
all it provided.

3) Updated snipwown.py to use weatherConditionMapper from self.provider
instead of using OWMToWeatherConditionMapper directly.
